[3] Update repo name
Update repo name on different files.
Move http service setup logic to the http_service.go file.
