+++ go.mod
@@ -1,11 +1,12 @@
-module github.com/sergionunezgo/gorest
+module github.com/sergionunezgo/goservice
 
 go 1.14
 
 require (
-	github.com/urfave/cli v1.22.4
-	go.uber.org/zap v1.15.0
-	github.com/pkg/errors v0.9.1
 	github.com/gorilla/mux v1.7.4
+	github.com/pkg/errors v0.9.1
+	github.com/sergionunezgo/gorest v0.0.0-20200728053535-34e7fe647f68
+	github.com/urfave/cli v1.22.4
 	github.com/urfave/negroni v1.0.0
+	go.uber.org/zap v1.15.0
 )
