+++ script/build.sh
@@ -2,4 +2,4 @@
 
 echo "running go build"
 
-CGO_ENABLED=0 go build -o build/gorest ./cmd/gorest
+CGO_ENABLED=0 go build -o build/goservice ./cmd/goservice
